# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dtos.ForgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
...
  dtos.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
      remember_me:
        description: RememberMe extends the iam_session cookie lifetime (SSO_SESSION_REMEMBER_TTL).
          API login and POST /oidc/login both honor this.
        type: boolean
      return_to:
        description: 'ReturnTo when set (OIDC browser flow): creates a session cookie
          and redirects here with 302 — no access_token in the response body.'
        type: string
      tenant_id:
        description: TenantID optional explicit tenant context.
        type: string
    required:
    - email
    - password
    type: object
  dtos.LoginResponse:
    properties:
      access_token:
        type: string
      active_tenant_id:
        type: string
      expires_in:
        description: access token lifetime (seconds)
        type: integer
      refresh_expires_in:
        description: refresh token lifetime (seconds)
        type: integer
      refresh_token:
        type: string
      token_type:
        type: string
    type: object
...
  dtos.Meta:
    description: Metadata for pagination
    properties:
      code:
        example: 400
        type: integer
      error_code:
        example: BAD_REQUEST
        type: string
      message:
        example: Request is invalid
        type: string
      page:
        example: 1
        type: integer
      page_size:
        example: 20
        type: integer
      total:
        example: 18
        type: integer
    type: object
...
  dtos.PublicFederationProviderResponse:
    properties:
      name:
        type: string
      provider:
        type: string
    type: object
  dtos.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  dtos.RegisterRequest:
    properties:
      email:
        example: user@example.com
        type: string
      first_name:
        maxLength: 100
        type: string
      last_name:
        maxLength: 100
        type: string
      password:
        example: secretpassword
        maxLength: 128
        minLength: 12
        type: string
      tenant_id:
        type: string
    required:
    - email
    - password
    type: object
  dtos.ResetPasswordRequest:
    properties:
      new_password:
        maxLength: 128
        minLength: 12
        type: string
      token:
        type: string
    required:
    - new_password
    - token
    type: object
  dtos.TenantSelectRequest:
    properties:
      remember_me:
        type: boolean
      selection_token:
        type: string
      tenant_id:
        type: string
    required:
    - selection_token
    - tenant_id
    type: object
  dtos.TenantSummary:
    properties:
      domain:
        type: string
      id:
        type: string
      name:
        type: string
      role:
        type: string
    type: object
  dtos.TenantSwitchRequest:
    properties:
      tenant_id:
        type: string
    required:
    - tenant_id
    type: object
...
  dtos.UpdateProfileRequest:
    properties:
      first_name:
        maxLength: 100
        minLength: 1
        type: string
      last_name:
        maxLength: 100
        type: string
    type: object
  dtos.UserResponse:
    properties:
      active_tenant_id:
        type: string
      created_at:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      first_name:
        type: string
      id:
        type: string
      is_platform_admin:
        type: boolean
      last_name:
        type: string
      mfa_enabled:
        type: boolean
      tenants:
        items:
          $ref: '#/definitions/dtos.TenantSummary'
        type: array
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
  description: GateForge IAM application API under /api/v1. The canonical public SDK
    contract (including root OIDC paths) lives at api/openapi.yaml in the monorepo.
  title: GateForge IAM API
  version: 0.1.0
paths:
...
  /federation/providers:
    get:
      parameters:
      - description: Tenant UUID (defaults to DEFAULT_TENANT_ID)
        in: query
        name: tenant_id
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/dtos.PublicFederationProviderResponse'
                type: array
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
      summary: List enabled federation sign-in methods for a tenant
      tags:
      - Auth
  /forgot-password:
    post:
      consumes:
      - application/json
      - application/json
      description: Always returns 200 so callers cannot tell whether the email exists
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dtos.RefreshTokenRequest'
      - description: Email
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dtos.ForgotPasswordRequest'
      produces:
      - application/json
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
      summary: Request a password reset
      tags:
      - Auth
      - Auth
...
  /login:
    post:
      consumes:
      - application/json
      description: Sets the same iam_session cookie as OIDC so /authorize recognizes
        the browser without a second login. If the user has MFA (TOTP) enabled, `data`
        is MFALoginChallengeResponse (mfa_ticket) instead of LoginResponse; complete
        login with POST /mfa/challenge/verify. When the user belongs to multiple tenants
        and no tenant context is provided, `data` is TenantSelectionResponse.
      parameters:
      - description: Credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dtos.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'When MFA is off: LoginResponse. When MFA is on: same envelope
            but `data` is MFALoginChallengeResponse (see top-level description).'
          schema:
            properties:
              data:
                $ref: '#/definitions/dtos.LoginResponse'
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
      summary: Login (dashboard/API) with email and password
      tags:
      - Auth
  /login/session:
    post:
      description: Reads the browser iam_session cookie set by federation callback
        and returns JWT tokens for the SPA dashboard. Used by GET /login/federation/complete.
      produces:
      - application/json
      responses:
        "200":
          description: When tenant is resolved. When multiple tenants apply, `data`
            is TenantSelectionResponse.
          schema:
            properties:
              data:
                $ref: '#/definitions/dtos.LoginResponse'
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
      summary: Exchange iam_session cookie for API tokens (after federated sign-in)
      tags:
      - Auth
  /logout:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
      security:
      - BearerAuth: []
      summary: Log out everywhere for the current user (SSO session + all refresh
        tokens)
      tags:
      - Auth
  /me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                $ref: '#/definitions/dtos.UserResponse'
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
      security:
      - BearerAuth: []
      summary: Current user profile
      tags:
      - Auth
    patch:
      consumes:
      - application/json
      parameters:
      - description: Profile fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dtos.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                $ref: '#/definitions/dtos.UserResponse'
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
      security:
      - BearerAuth: []
      summary: Update current user profile
      tags:
      - Auth
  /me/tenants:
    get:
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/dtos.TenantSummary'
                type: array
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
      security:
      - BearerAuth: []
      summary: List tenants the current user can access
      tags:
      - Auth
...
  /refresh:
    post:
      consumes:
      - application/json
      - application/json
      description: Always returns 200 so callers cannot tell whether the email exists
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dtos.RefreshTokenRequest'
      - description: Email
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dtos.ForgotPasswordRequest'
      produces:
      - application/json
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
      summary: Request a password reset
      tags:
      - Auth
      - Auth
  /register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dtos.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                $ref: '#/definitions/dtos.UserResponse'
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
      summary: Register a new user
      tags:
      - Auth
  /reset-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Token and new password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dtos.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
      summary: Reset a password with a one-time token
      tags:
      - Auth
  /tenants/select:
    post:
      consumes:
      - application/json
      parameters:
      - description: Selection token and tenant
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dtos.TenantSelectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                $ref: '#/definitions/dtos.LoginResponse'
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
      summary: Complete login by selecting a tenant
      tags:
      - Auth
  /tenants/switch:
    post:
      consumes:
      - application/json
      parameters:
      - description: Target tenant
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dtos.TenantSwitchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                $ref: '#/definitions/dtos.LoginResponse'
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
      security:
      - BearerAuth: []
      summary: Switch active tenant and re-issue tokens
      tags:
      - Auth
...
schemes:
- http
- https
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    description: Bearer Token Authentication. Use "Bearer {token}" as the value.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"